requirements.txt
atlassian-python-api
zdesk
